% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcmdrPlugin.GWRM.r
\name{generalizedWaringModel}
\alias{generalizedWaringModel}
\title{Generalized Waring Regression Model
}
\usage{
generalizedWaringModel()
}
\value{
a new model in the model list
}
\description{
Generalized Waring Regression Model
Using the selected data show a window asking for the GWRM parameters
}

