% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcmdrPlugin.GWRM.r
\name{gwrmAddObservationStatistics}
\alias{gwrmAddObservationStatistics}
\title{Add observationStatisticsGWRM
Using gw model active add observations data}
\usage{
gwrmAddObservationStatistics()
}
\value{
Residuals plot
}
\description{
Add observationStatisticsGWRM
Using gw model active add observations data
}

