% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcmdrPlugin.GWRM.r
\name{gwrmPartVar}
\alias{gwrmPartVar}
\title{Generalized Waring Regression Model Partition Varianze
Using the selected gw model show a window asking for de Partition Varianze values}
\usage{
gwrmPartVar()
}
\value{
Partition Varianze of the model
}
\description{
Generalized Waring Regression Model Partition Varianze
Using the selected gw model show a window asking for de Partition Varianze values
}

