% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.GWRM.r
\name{RcmdrPlugin.Utility}
\alias{RcmdrPlugin.Utility}
\alias{extractNewData}
\alias{listGeneralizedWaringModels}
\alias{activeModelGW}
\alias{showTable}
\alias{gwrmPartVar}
\alias{gwrmResiduals}
\alias{gwrmAddObservationStatistics}
\title{GWRM Plug-In Utility Functions}
\usage{
extractNewData(nrows, colNames)

listGeneralizedWaringModels(envir = .GlobalEnv, ...)

activeModelGW(envir = .GlobalEnv, ...)

showTable(...)

gwrmPartVar()

gwrmResiduals()

gwrmAddObservationStatistics()
}
\arguments{
\item{nrows}{the new data number rows}

\item{colNames}{the names of the covariates}

\item{envir}{the environments inside search.}

\item{...}{further arguments.}
}
\description{
GWRM Plug-In Utility Functions
}
