% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcmdrPlugin.GWRM.r
\name{generalizedWaringModel}
\alias{generalizedWaringModel}
\title{Rcmdr Generalized Waring Regression Model Dialog}
\usage{
generalizedWaringModel()
}
\description{
This dialog is used to specify a generalized Waring regression model to be fit by the gw function.
}
\details{
This dialog box is similar to that of the Rcmdr generalized linear model. The main differences are the "k parameter" box, which allows you to introduce a fixed value for k (it is optional) and the absence of the "Family" box.
}
\author{
Antonio Jose Saez-Castillo, Maria Jose Olmo-Jimenez and Silverio Vilchez-Lopez.
}
\seealso{
gw, glm, Rcmdr
}

