% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcmdrPlugin.GWRM.r
\name{gwrmAddObservationStatistics}
\alias{gwrmAddObservationStatistics}
\title{Add observationStatisticsGWRM
Using gw model active add observations data
The same Rcmdr function with two bug solved}
\usage{
gwrmAddObservationStatistics()
}
\value{
nothing
}
\description{
Add observationStatisticsGWRM
Using gw model active add observations data
The same Rcmdr function with two bug solved
}

