% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcmdrPlugin.GWRM.r
\name{gwrmPartVar}
\alias{gwrmPartVar}
\title{Partition of the variance in the generalized Waring regression model
}
\usage{
gwrmPartVar()
}
\description{
Rcmdr menu for the partition of the variance in the selected generalized Waring regression model.
}
\details{
If desired you can enter new data to obtain the partition of the variance in the selected gw model.
}
\author{
Antonio Jose Saez-Castillo, Maria Jose Olmo-Jimenez and Silverio Vilchez-Lopez.
}
\seealso{
partvar, gw
}

