% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcmdrPlugin.GWRM.r
\name{gwrmResiduals}
\alias{gwrmResiduals}
\title{Generalized Waring Regression Model Residuals
Using the selected gw model show a window asking for the residuals function parameters}
\usage{
gwrmResiduals()
}
\value{
Residuals plot
}
\description{
Generalized Waring Regression Model Residuals
Using the selected gw model show a window asking for the residuals function parameters
}

